% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotIncidence.R
\name{plotIncidence}
\alias{plotIncidence}
\title{Plot incidence time series}
\usage{
plotIncidence(x, dates, bin = 7, fill.by = NULL, split.by = NULL,
  shade.by = NULL, start.at = NULL, stop.at = NULL)
}
\arguments{
\item{x}{a data.frame containing the data to be plotted.}

\item{dates}{a character string or an integer indicating which column contains the dates to use}

\item{bin}{an integer indicating the size of the time window to use for the incidence computation}

\item{fill.by}{a character string or an integer indicating which column to use to color the bars}

\item{split.by}{a character string or an integer indicating which column to use to split the charts}

\item{shade.by}{a character string or an integer indicating which column to use to shade the bars}

\item{start.at}{a starting date for the incidence; defaults to the earliest date of the dataset}

\item{start.at}{a stoping date for the incidence; defaults to the latest date of the dataset}
}
\description{
This function provides advanced plotting facilities for incidence time series using ggplot2
}
\examples{
data("toy")
head(toy)
plotIncidence(toy, dates="dateOfOnset")
}

