% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docToy.R
\name{toy}
\alias{toy}
\title{Toy outbreak dataset}
\format{A data frame with 187 rows and 3 variables}
\source{
Michael Hohle's dataset Hagelloch
}
\description{
This is a simplified version of Michael Hohle's dataset Hagelloch
}
\details{
\itemize{
  \item dateOfOnset a vector of dates of symptom onset in Date format
  \item gender the gender of the patient
  \item complication an indication of possible clinical complications
}
}

