% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/doc.R
\docType{package}
\name{vizIncidence}
\alias{vizIncidence}
\alias{vizIncidence-package}
\title{plotIncidence: A package for plotting incidence data.}
\description{
This package provides functions for computing and plotting incidence data from outbreak line lists.
}
\section{Available functions include}{

 \itemize{
 \item{plotIncidence}{Plot incidence time series using histograms}
 \item{mapIncidence}{Visualise incidence in space and time using animations}
}
}

