% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docToy.R, R/docZombie.R
\name{zombie_outbreak}
\alias{zombie_outbreak}
\title{#' #' Zombie outbreak dataset}
\format{A data frame with 154013 rows and 4 variables}
\description{
This is a dataset simulating an outbreak with geo-localisation of the cases (6 localities)

This is a dataset simulating an outbreak with geo-localisation of the cases (6 localities)
}
\details{
\itemize{
  \item x.coord a vector of x coordinates
  \item y.coord a vector of y coordinates
  \item date a vector of dates of recorded cases in Date format
  \item gender the gender of the patient
}

\itemize{
  \item x.coord a vector of x coordinates
  \item y.coord a vector of y coordinates
  \item date a vector of dates of recorded cases in Date format
  \item gender the gender of the patient
}
}

